# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ANONYMIKETECH WhatsApp Bot Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bots/save-session:
    post:
      tags: [bots]
      operationId: saveSession
      summary: Save a WhatsApp bot session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveSessionRequest'
      responses:
        '201':
          description: Bot session saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /bots/my-bots:
    get:
      tags: [bots]
      operationId: getMyBots
      summary: Get all bots for the current user
      responses:
        '200':
          description: List of bots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotsListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /bots/start-bot:
    post:
      tags: [bots]
      operationId: startBot
      summary: Start a bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BotActionRequest'
      responses:
        '200':
          description: Bot started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotResponse'
        '400':
          description: Insufficient coins or bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Bot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /bots/stop-bot:
    post:
      tags: [bots]
      operationId: stopBot
      summary: Stop a bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BotActionRequest'
      responses:
        '200':
          description: Bot stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BotResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Bot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    Bot:
      type: object
      required: [id, name, status, coins, expiresAt, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [running, stopped]
        coins:
          type: number
        expiresAt:
          type: [string, 'null']
          format: date-time
        createdAt:
          type: string
          format: date-time
    BotResponse:
      type: object
      required: [bot]
      properties:
        bot:
          $ref: '#/components/schemas/Bot'
    BotsListResponse:
      type: object
      required: [bots]
      properties:
        bots:
          type: array
          items:
            $ref: '#/components/schemas/Bot'
    SaveSessionRequest:
      type: object
      required: [name, sessionId]
      properties:
        name:
          type: string
          minLength: 1
        sessionId:
          type: string
          minLength: 1
    BotActionRequest:
      type: object
      required: [botId]
      properties:
        botId:
          type: string
          minLength: 1
...