# lib/api-zod/src/generated/types/botActionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ANONYMIKETECH WhatsApp Bot Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface BotActionRequest {
  /** @minLength 1 */
  botId: string;
}