# lib/api-zod/src/generated/types/botResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ANONYMIKETECH WhatsApp Bot Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Bot } from "./bot";

export interface BotResponse {
  bot: Bot;
}