# lib/api-zod/src/generated/types/saveSessionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ANONYMIKETECH WhatsApp Bot Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveSessionRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  sessionId: string;
}